(#130) Move ytLinkId to KGBotka.Friday
